starting on general optimiser framework
created a more general `Node` class, which we can optimise on. Created an example of this in the examples folder. Also, to start with, I've created a simple brute force algorithm,
which just iterates across the whole configuration space. I create a backend folder, where I think we should work on the seperate parsers.